Agrego product details + alerta de confirmación en eliminación

+++ frontend/src/hooks/useDeleteProduct.ts
@@ -1,31 +1,49 @@
+import { useNavigate } from "react-router-dom";
 import toast from "react-hot-toast";
+import Swal from "sweetalert2";
 
-export function useDeleteProduct(dispatch: any, id: number) {
+export function useDeleteProduct(dispatch: any, id: number | undefined | null) {
+  const navigate = useNavigate();
+
   const handleDelete = async () => {
-    const action = {
-      type: "DELETE PRODUCT",
-      payload: id,
-    };
+    //Dynamic alert to confirm deleting
+    const confirmResult = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "green",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it",
+      cancelButtonText: "Cancel",
+      background: "#202020",
+      color: "#eee",
+    });
 
-    const waitingToast = toast.loading("Deleting product...");
+    if (confirmResult.isConfirmed) {
+      const action = {
+        type: "DELETE PRODUCT",
+        payload: id,
+      };
 
-    try {
-      const res = await fetch(`http://localhost:3000/products/${id}`, {
-        method: "DELETE",
-      });
+      const waitingToast = toast.loading("Deleting product...");
 
-      if (res.ok) {
+      try {
+        const res = await fetch(`http://localhost:3000/products/${id}`, {
+          method: "DELETE",
+        });
+
+        if (res.ok) {
+          toast.dismiss(waitingToast);
+          dispatch(action);
+          toast.success("Product deleted!");
+          navigate("/");
+        }
+      } catch (error) {
         toast.dismiss(waitingToast);
-        dispatch(action);
-        toast.success("Product deleted!");
-      } else {
-        toast.dismiss(waitingToast);
-        toast.error("Error deleting product");
+        console.error(error);
+        toast.error("Error deleting product.");
       }
-    } catch (error) {
-      toast.dismiss(waitingToast);
-      console.error(error);
-      toast.error("Error deleting product.");
     }
   };
